Updated and corrected the wf3cte.rst file.
Removed background documentation on the CTE algorithm which does not
belong in this source file.  Replaced with URLs for handy reference,
particularly so the documentation is not duplicated in multiple
locations, and the latest information is available for viewing.
Removed support for incorrect items from the source: ability to
set the output filename and ability for the code to accept an
association table as input.  Removal of the output filename modifies
the API.