update embedsub.py
The previous code created a new fits file and only appended the sci, err, and dq data and header extensions. This was causing problems down the line with astrodrizzle because the D2MIARR extensions were no longer a part of the file (which is a requirement for astrodrizzle). The code now makes a copy of the input file so that all the extensions get carried over to the embedded output file.